Adjust qrexec_timeout to 600 seconds
The default timeout of 60 seconds for starting domains is to low for slower storage or busy systems.
The timeout causes the killing of the starting domain.
This is fatal during the installation process.
I have experienced a lot of different problems causing different effects, such as
* domain starts, but to application won't open
* network isn't working correctly

During the first start of a domain some initialization is happening. Would this initialization process be interrupted the domain will behave unexpected.

This is a hotfix.
A long-term solution would be, when the expiring timeout wouldn't kill the domain. It should inform the customer and let them decide whether to kill the domain or to be patient.